Version 1.0.0. (#97)
* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Konnexions Public License (KX-PL).

* Version 1.0.0.

* Version 1.0.0.

* Version 1.0.0.

* Update .travis.yml

+++ src_erlang/scripts/run_bench_oranif.bat
@@ -53,21 +53,23 @@ echo ===========================================================================
 if NOT ["%ORA_BENCH_MULTIPLE_RUN%"] == ["true"] (
     call src_java\scripts\run_gradle
     if %ERRORLEVEL% NEQ 0 (
+        echo Processing of the script was aborted, error code=%ERRORLEVEL%
         exit %ERRORLEVEL%
     )
 
     java -cp "%ORA_BENCH_JAVA_CLASSPATH%" ch.konnexions.orabench.OraBench setup_erlang
     if %ERRORLEVEL% NEQ 0 (
+        echo Processing of the script was aborted, error code=%ERRORLEVEL%
         exit %ERRORLEVEL%
     )
 
     cd src_erlang
     
-    if EXIST _build\ 
-        rd /Q/S _build 
+    if EXIST _build\ rd /Q/S _build 
         
     call rebar3 escriptize
     if %ERRORLEVEL% NEQ 0 (
+        echo Processing of the script was aborted, error code=%ERRORLEVEL%
         exit %ERRORLEVEL%
     )
 
@@ -76,6 +78,7 @@ if NOT ["%ORA_BENCH_MULTIPLE_RUN%"] == ["true"] (
 
 src_erlang\_build\default\bin\orabench priv\properties\ora_bench_erlang.properties oranif
 if %ERRORLEVEL% NEQ 0 (
+    echo Processing of the script was aborted, error code=%ERRORLEVEL%
     exit %ERRORLEVEL%
 )
 
@@ -84,5 +87,3 @@ echo:| TIME
 echo --------------------------------------------------------------------------------
 echo End   %0
 echo ================================================================================
-
-exit %ERRORLEVEL%
